@@ -10,12 +10,12 @@ namespace Penumbra.GameData.DataContainers;
 
 /// <summary> A dictionary that maps full item types to lists of all corresponding items. </summary>
 public sealed class ItemsByType(DalamudPluginInterface pi, Logger log, IDataManager dataManager)
-    : DataSharer<IReadOnlyList<IReadOnlyList<PseudoEquipItem>>>(pi, log, "ItemsByType", dataManager.Language, 1,
+    : DataSharer<IReadOnlyList<IReadOnlyList<PseudoEquipItem>>>(pi, log, "ItemsByType", dataManager.Language, 2,
             () => CreateItems(dataManager)),
         IReadOnlyDictionary<FullEquipType, IReadOnlyList<EquipItem>>
 {
     /// <summary> Create the data. </summary>
-    private static IReadOnlyList<IReadOnlyList<PseudoEquipItem>> CreateItems(IDataManager dataManager)
+    private static IReadOnlyList<PseudoEquipItem>[] CreateItems(IDataManager dataManager)
     {
         var tmp = Enum.GetValues<FullEquipType>().Select(_ => new List<EquipItem>(1024)).ToArray();
 
@@ -95,7 +95,7 @@ public sealed class ItemsByType(DalamudPluginInterface pi, Logger log, IDataMana
             return true;
         }
 
-        value = Array.Empty<EquipItem>();
+        value = [];
         return false;
     }
 
