@@ -1,7 +1,8 @@
-using Dalamud.Plugin.Services;
+﻿using Dalamud.Plugin.Services;
 using OtterGui.Services;
 using Penumbra.GameData.Data;
 using Penumbra.GameData.Enums;
+using Penumbra.GameData.Interop;
 
 namespace Penumbra.GameData.Actors;
 
@@ -12,9 +13,12 @@ public delegate short CutsceneResolver(ushort index);
 public sealed class ActorManager : ActorIdentifierFactory, IDisposable, IAsyncService
 {
     /// <summary> The names used for NPC types. </summary>
-    public readonly  NameDicts     Data;
+    public readonly NameDicts Data;
+
     private readonly ActorResolver _resolver;
     private readonly IClientState  _clientState;
+    private          uint          _homeWorld;
+
 
     /// <summary> Waits for the NameDicts to be ready. </summary>
     public Task Awaiter
@@ -29,22 +33,33 @@ public sealed class ActorManager : ActorIdentifierFactory, IDisposable, IAsyncSe
     {
         if (ActorIdentifierExtensions.Manager == this)
             ActorIdentifierExtensions.Manager = null;
+        _clientState.Login  -= OnLogin;
+        _clientState.Logout -= OnLogout;
     }
 
     public ActorManager(NameDicts data,
         IFramework framework,
-        IObjectTable objects,
+        ObjectManager objects,
         IClientState clientState,
         IGameGui gameGui,
         CutsceneResolver toParentIdx)
         : base(objects, framework, data, toParentIdx)
     {
-        Data                              =   data;
-        _clientState                      =   clientState;
-        _resolver                         =   new ActorResolver(gameGui, objects, clientState);
+        Data         = data;
+        _clientState = clientState;
+        _resolver    = new ActorResolver(gameGui, objects, clientState);
         // Set the static manager if it is unset.
         ActorIdentifierExtensions.Manager ??= this;
+        _clientState.Login                +=  OnLogin;
+        _clientState.Logout               +=  OnLogout;
+        _homeWorld                        =   _clientState.LocalPlayer?.HomeWorld.Id ?? 0;
     }
+
+    private void OnLogin()
+        => _homeWorld = _clientState.LocalPlayer?.HomeWorld.Id ?? _homeWorld;
+
+    private void OnLogout()
+        => _homeWorld = 0;
 
     /// <inheritdoc cref="ActorResolver.GetCurrentPlayer"/>
     public ActorIdentifier GetCurrentPlayer()
@@ -79,18 +94,18 @@ public sealed class ActorManager : ActorIdentifierFactory, IDisposable, IAsyncSe
     {
         return id.Type switch
         {
-            IdentifierType.Player => id.HomeWorld.Id != _clientState.LocalPlayer?.HomeWorld.Id
+            IdentifierType.Player => id.HomeWorld.Id != _homeWorld
                 ? $"{id.PlayerName} ({Data.ToWorldName(id.HomeWorld)})"
                 : id.PlayerName.ToString(),
             IdentifierType.Retainer => $"{id.PlayerName}{id.Retainer switch
             {
-                ActorIdentifier.RetainerType.Bell      => " (Bell)",
-                ActorIdentifier.RetainerType.Mannequin => " (Mannequin)",
-                _                                      => " (Retainer)",
+                ActorIdentifier.RetainerType.Bell      => " (传唤铃雇员)",
+                ActorIdentifier.RetainerType.Mannequin => " (服装模特)",
+                _                                      => " (雇员)",
             }}",
-            IdentifierType.Owned => id.HomeWorld.Id != _clientState.LocalPlayer?.HomeWorld.Id
-                ? $"{id.PlayerName} ({Data.ToWorldName(id.HomeWorld)})'s {Data.ToName(id.Kind, id.DataId)}"
-                : $"{id.PlayerName}s {Data.ToName(id.Kind,                                     id.DataId)}",
+            IdentifierType.Owned => id.HomeWorld.Id != _homeWorld
+                ? $"{id.PlayerName} ({Data.ToWorldName(id.HomeWorld)})的{Data.ToName(id.Kind, id.DataId)}"
+                : $"{id.PlayerName}的{Data.ToName(id.Kind,                                     id.DataId)}",
             IdentifierType.Special => ((ScreenActor)id.Index.Index).ToName(),
             IdentifierType.Npc =>
                 id.Index == ushort.MaxValue
