@@ -3,7 +3,7 @@ namespace Penumbra.GameData.Data;
 internal static class NpcNames
 {
     /// <summary> Generate an outer list indexed by the BNPC ID, with the inner lists collecting all known BNPC Name IDs associated with that BNPC ID. </summary>
-    /// <remarks> Generated from https://api.ffxivteamcraft.com/gubal on 2024-01-17. </remarks>
+    /// <remarks> Generated from https://api.ffxivteamcraft.com/gubal on 2024-03-20. </remarks>
     public static IReadOnlyList<IReadOnlyList<uint>> CreateNames()
         => new IReadOnlyList<uint>[]
         {
@@ -5537,7 +5537,7 @@ internal static class NpcNames
             [ 4687 ],
             [ 3646 ],
             [ 4688 ],
-            [ ],
+            [ 4689 ],
             [ 4690 ],
             [ 4797 ],
             [ 4701 ],
@@ -9156,7 +9156,7 @@ internal static class NpcNames
             [ 7504 ],
             [ 7503 ],
             [ 7523 ],
-            [ ],
+            [ 7523 ],
             [ 7523 ],
             [ 7523 ],
             [ 7524 ],
@@ -9515,16 +9515,16 @@ internal static class NpcNames
             [ 7880 ],
             [ 7881 ],
             [ 7882 ],
-            [ ],
+            [ 7895 ],
             [ 7884 ],
             [ 7885 ],
             [ 7886 ],
             [ 7887 ],
             [ 7888 ],
-            [ ],
-            [ ],
-            [ ],
-            [ ],
+            [ 7896 ],
+            [ 7897 ],
+            [ 7898 ],
+            [ 7895 ],
             [ 6154 ],
             [ 7715 ],
             [ 6196 ],
@@ -9602,7 +9602,7 @@ internal static class NpcNames
             [ 7248 ],
             [ 7765 ],
             [ 7766 ],
-            [ ],
+            [ 7893 ],
             [ 7751 ],
             [ 7752 ],
             [ 7248 ],
@@ -9659,8 +9659,8 @@ internal static class NpcNames
             [ 7858 ],
             [ 7861 ],
             [ 7862 ],
-            [ ],
-            [ ],
+            [ 108 ],
+            [ 7895 ],
             [ 7861 ],
             [ 7664 ],
             [ 7919 ],
@@ -9748,7 +9748,7 @@ internal static class NpcNames
             [ 108 ],
             [ 108 ],
             [ 108 ],
-            [ ],
+            [ 108 ],
             [ 7879 ],
             [ 108 ],
             [ ],
@@ -9816,11 +9816,11 @@ internal static class NpcNames
             [ 8002 ],
             [ 8003 ],
             [ 8004 ],
-            [ ],
-            [ ],
-            [ ],
-            [ ],
-            [ ],
+            [ 7896 ],
+            [ 7896 ],
+            [ 7896 ],
+            [ 7894 ],
+            [ 1649 ],
             [ 8922 ],
             [ 7930 ],
             [ 7930 ],
@@ -10053,7 +10053,7 @@ internal static class NpcNames
             [ 8127 ],
             [ 7155 ],
             [ 8068 ],
-            [ ],
+            [ 108 ],
             [ 7954 ],
             [ 108 ],
             [ 8066 ],
@@ -16320,7 +16320,7 @@ internal static class NpcNames
             [ 12575 ],
             [ ],
             [ 12577 ],
-            [ ],
+            [ 12578 ],
             [ 12579 ],
             [ 12580 ],
             [ ],
