@@ -10,7 +10,7 @@ namespace Penumbra.GameData.DataContainers;
 
 /// <summary> A dictionary mapping a single BNpcBaseId to a list of BNpcNameIds, using the data generated by the Gubal API (cf. <see cref="NpcNames"/>). </summary>
 public sealed class DictBNpcNames(DalamudPluginInterface pluginInterface, Logger log)
-    : DataSharer<IReadOnlyList<IReadOnlyList<uint>>>(pluginInterface, log, "BNpcNameDict", ClientLanguage.English, 2, NpcNames.CreateNames),
+    : DataSharer<IReadOnlyList<IReadOnlyList<uint>>>(pluginInterface, log, "BNpcNameDict", ClientLanguage.English, 3, NpcNames.CreateNames),
         IReadOnlyDictionary<BNpcId, IReadOnlyList<BNpcNameId>>
 {
 
