@@ -1,3 +1,4 @@
+using System.Collections.Frozen;
 using Dalamud.Plugin;
 using Dalamud.Plugin.Services;
 using OtterGui.Log;
@@ -8,12 +9,12 @@ namespace Penumbra.GameData.DataContainers;
 
 /// <summary> A dictionary mapping ItemIds to all primary model items. This requires ItemsByType to be finished. </summary>
 public sealed class ItemsPrimaryModel(DalamudPluginInterface pi, Logger log, IDataManager gameData, ItemsByType items)
-    : ItemDictionary(pi, log, "ItemDictPrimary", gameData.Language, 2, () => CreateMainItems(items), items.Awaiter)
+    : ItemDictionary(pi, log, "ItemDictPrimary", gameData.Language, 3, () => CreateMainItems(items), items.Awaiter)
 {
     /// <summary> Create data by taking only the primary models for all items. </summary>
     private static IReadOnlyDictionary<uint, PseudoEquipItem> CreateMainItems(ItemsByType items)
     {
-        var dict = new Dictionary<uint, PseudoEquipItem>(1024 * 4);
+        var dict = new Dictionary<uint, PseudoEquipItem>(1024 * 16);
         foreach (var type in Enum.GetValues<FullEquipType>().Where(v => !FullEquipTypeExtensions.OffhandTypes.Contains(v)))
         {
             var list = items.Value[(int)type];
@@ -29,8 +30,6 @@ public sealed class ItemsPrimaryModel(DalamudPluginInterface pi, Logger log, IDa
             }
         }
 
-        // TODO: FrozenDictionary.
-        dict.TrimExcess();
-        return dict;
+        return dict.ToFrozenDictionary();
     }
 }
