@@ -1,3 +1,4 @@
+using System.Collections.Frozen;
 using Dalamud.Plugin;
 using Dalamud.Plugin.Services;
 using OtterGui.Log;
@@ -8,12 +9,12 @@ namespace Penumbra.GameData.DataContainers;
 
 /// <summary> A dictionary mapping ItemIds to all secondary model items (offhands). This requires ItemsByType to be finished. </summary>
 public sealed class ItemsSecondaryModel(DalamudPluginInterface pi, Logger log, IDataManager gameData, ItemsByType items)
-    : ItemDictionary(pi, log, "ItemDictSecondary", gameData.Language, 1, () => CreateOffhands(items), items.Awaiter)
+    : ItemDictionary(pi, log, "ItemDictSecondary", gameData.Language, 2, () => CreateOffhands(items), items.Awaiter)
 {
     /// <summary> Create data by taking only the secondary models for all items. </summary>
     private static IReadOnlyDictionary<uint, PseudoEquipItem> CreateOffhands(ItemsByType items)
     {
-        var dict = new Dictionary<uint, PseudoEquipItem>(1024);
+        var dict = new Dictionary<uint, PseudoEquipItem>(1024 * 4);
         foreach (var type in FullEquipTypeExtensions.OffhandTypes)
         {
             var list = items.Value[(int)type];
@@ -21,8 +22,6 @@ public sealed class ItemsSecondaryModel(DalamudPluginInterface pi, Logger log, I
                 dict.TryAdd((uint)item.Item2, item);
         }
 
-        // TODO: FrozenDictionary
-        dict.TrimExcess();
-        return dict;
+        return dict.ToFrozenDictionary();
     }
 }
