@@ -103,42 +103,42 @@ public static class CustomizationExtensions
     public static string ToDefaultName(this CustomizeIndex customizeIndex)
         => customizeIndex switch
         {
-            CustomizeIndex.Race              => "Race",
-            CustomizeIndex.Gender            => "Gender",
-            CustomizeIndex.BodyType          => "Body Type",
-            CustomizeIndex.Height            => "Height",
-            CustomizeIndex.Clan              => "Clan",
-            CustomizeIndex.Face              => "Head Style",
-            CustomizeIndex.Hairstyle         => "Hair Style",
-            CustomizeIndex.Highlights        => "Enable Highlights",
-            CustomizeIndex.SkinColor         => "Skin Color",
-            CustomizeIndex.EyeColorRight     => "Left Eye", // inverted due to compatibility bug.
-            CustomizeIndex.HairColor         => "Hair Color",
-            CustomizeIndex.HighlightsColor   => "Highlights Color",
-            CustomizeIndex.TattooColor       => "Tattoo Color",
-            CustomizeIndex.Eyebrows          => "Eyebrow Style",
-            CustomizeIndex.EyeColorLeft      => "Right Eye", // inverted due to compatibility bug.
-            CustomizeIndex.EyeShape          => "Small Pupils",
-            CustomizeIndex.Nose              => "Nose Style",
-            CustomizeIndex.Jaw               => "Jaw Style",
-            CustomizeIndex.Mouth             => "Mouth Style",
-            CustomizeIndex.MuscleMass        => "Muscle Tone",
-            CustomizeIndex.TailShape         => "Tail Shape",
-            CustomizeIndex.BustSize          => "Bust Size",
-            CustomizeIndex.FacePaint         => "Face Paint",
-            CustomizeIndex.FacePaintColor    => "Face Paint Color",
-            CustomizeIndex.LipColor          => "Lip Color",
-            CustomizeIndex.FacialFeature1    => "Facial Feature 1",
-            CustomizeIndex.FacialFeature2    => "Facial Feature 2",
-            CustomizeIndex.FacialFeature3    => "Facial Feature 3",
-            CustomizeIndex.FacialFeature4    => "Facial Feature 4",
-            CustomizeIndex.FacialFeature5    => "Facial Feature 5",
-            CustomizeIndex.FacialFeature6    => "Facial Feature 6",
-            CustomizeIndex.FacialFeature7    => "Facial Feature 7",
-            CustomizeIndex.LegacyTattoo      => "Legacy Tattoo",
-            CustomizeIndex.SmallIris         => "Small Iris",
-            CustomizeIndex.Lipstick          => "Enable Lipstick",
-            CustomizeIndex.FacePaintReversed => "Reverse Face Paint",
+            CustomizeIndex.Race              => "种族",
+            CustomizeIndex.Gender            => "性别",
+            CustomizeIndex.BodyType          => "身型",
+            CustomizeIndex.Height            => "身高",
+            CustomizeIndex.Clan              => "部族",
+            CustomizeIndex.Face              => "脸型",
+            CustomizeIndex.Hairstyle         => "发型",
+            CustomizeIndex.Highlights        => "启用挑染",
+            CustomizeIndex.SkinColor         => "肤色",
+            CustomizeIndex.EyeColorRight     => "左眼颜色", // inverted due to compatibility bug.
+            CustomizeIndex.HairColor         => "发色",
+            CustomizeIndex.HighlightsColor   => "挑染颜色",
+            CustomizeIndex.TattooColor       => "纹身颜色",
+            CustomizeIndex.Eyebrows          => "眉形",
+            CustomizeIndex.EyeColorLeft      => "右眼颜色", // inverted due to compatibility bug.
+            CustomizeIndex.EyeShape          => "小瞳孔",
+            CustomizeIndex.Nose              => "鼻型",
+            CustomizeIndex.Jaw               => "脸部轮廓",
+            CustomizeIndex.Mouth             => "嘴型",
+            CustomizeIndex.MuscleMass        => "肌肉",
+            CustomizeIndex.TailShape         => "尾巴形状",
+            CustomizeIndex.BustSize          => "胸围",
+            CustomizeIndex.FacePaint         => "面妆",
+            CustomizeIndex.FacePaintColor    => "面妆颜色",
+            CustomizeIndex.LipColor          => "唇色",
+            CustomizeIndex.FacialFeature1    => "黑痣与伤痕等 1",
+            CustomizeIndex.FacialFeature2    => "黑痣与伤痕等 2",
+            CustomizeIndex.FacialFeature3    => "黑痣与伤痕等 3",
+            CustomizeIndex.FacialFeature4    => "黑痣与伤痕等 4",
+            CustomizeIndex.FacialFeature5    => "黑痣与伤痕等 5",
+            CustomizeIndex.FacialFeature6    => "黑痣与伤痕等 6",
+            CustomizeIndex.FacialFeature7    => "黑痣与伤痕等 7",
+            CustomizeIndex.LegacyTattoo      => "遗产纹身",
+            CustomizeIndex.SmallIris         => "较小眼瞳",
+            CustomizeIndex.Lipstick          => "启用唇色",
+            CustomizeIndex.FacePaintReversed => "反转面妆",
             _                                => string.Empty,
         };
 }
