@@ -1,3 +1,5 @@
+using System.Collections.Frozen;
+
 namespace Penumbra.GameData.Enums;
 
 /// <summary> A grouping of files representing specific types for the game, not only used by extension. </summary>
@@ -21,25 +23,24 @@ public enum FileType : byte
 public static partial class Names
 {
     /// <summary> Dictionary to convert extensions to grouped FileType. </summary>
-    // TODO: FrozenDictionary
-    public static readonly Dictionary<string, FileType> ExtensionToFileType = new()
-    {
-        { ".mdl", FileType.Model },
-        { ".tex", FileType.Texture },
-        { ".mtrl", FileType.Material },
-        { ".atex", FileType.Animation },
-        { ".avfx", FileType.Vfx },
-        { ".scd", FileType.Sound },
-        { ".imc", FileType.Imc },
-        { ".pap", FileType.Pap },
-        { ".eqp", FileType.MetaInfo },
-        { ".eqdp", FileType.MetaInfo },
-        { ".est", FileType.MetaInfo },
-        { ".exd", FileType.MetaInfo },
-        { ".exh", FileType.MetaInfo },
-        { ".shpk", FileType.Shader },
-        { ".shcd", FileType.Shader },
-        { ".fdt", FileType.Font },
-        { ".envb", FileType.Environment },
-    };
+    public static readonly IReadOnlyDictionary<string, FileType> ExtensionToFileType = FrozenDictionary.ToFrozenDictionary(
+    [
+        new KeyValuePair<string, FileType>(".mdl",  FileType.Model),
+        new KeyValuePair<string, FileType>(".tex",  FileType.Texture),
+        new KeyValuePair<string, FileType>(".mtrl", FileType.Material),
+        new KeyValuePair<string, FileType>(".atex", FileType.Animation),
+        new KeyValuePair<string, FileType>(".avfx", FileType.Vfx),
+        new KeyValuePair<string, FileType>(".scd",  FileType.Sound),
+        new KeyValuePair<string, FileType>(".imc",  FileType.Imc),
+        new KeyValuePair<string, FileType>(".pap",  FileType.Pap),
+        new KeyValuePair<string, FileType>(".eqp",  FileType.MetaInfo),
+        new KeyValuePair<string, FileType>(".eqdp", FileType.MetaInfo),
+        new KeyValuePair<string, FileType>(".est",  FileType.MetaInfo),
+        new KeyValuePair<string, FileType>(".exd",  FileType.MetaInfo),
+        new KeyValuePair<string, FileType>(".exh",  FileType.MetaInfo),
+        new KeyValuePair<string, FileType>(".shpk", FileType.Shader),
+        new KeyValuePair<string, FileType>(".shcd", FileType.Shader),
+        new KeyValuePair<string, FileType>(".fdt",  FileType.Font),
+        new KeyValuePair<string, FileType>(".envb", FileType.Environment),
+    ]);
 }
