@@ -1,4 +1,4 @@
-using Lumina.Excel.GeneratedSheets;
+﻿using Lumina.Excel.GeneratedSheets;
 
 namespace Penumbra.GameData.Enums;
 
@@ -181,63 +181,63 @@ public static class FullEquipTypeExtensions
             FullEquipType.Ears            => EquipSlot.Ears.ToName(),
             FullEquipType.Neck            => EquipSlot.Neck.ToName(),
             FullEquipType.Wrists          => EquipSlot.Wrists.ToName(),
-            FullEquipType.Finger          => "Ring",
-            FullEquipType.Fists           => "Fist Weapon",
-            FullEquipType.FistsOff        => "Fist Weapon (Offhand)",
-            FullEquipType.Sword           => "Sword",
-            FullEquipType.Axe             => "Axe",
-            FullEquipType.Bow             => "Bow",
-            FullEquipType.BowOff          => "Quiver",
-            FullEquipType.Lance           => "Lance",
-            FullEquipType.Staff           => "Staff",
-            FullEquipType.Wand            => "Mace",
-            FullEquipType.Book            => "Book",
-            FullEquipType.Daggers         => "Dagger",
-            FullEquipType.DaggersOff      => "Dagger (Offhand)",
-            FullEquipType.Broadsword      => "Broadsword",
-            FullEquipType.Gun             => "Gun",
-            FullEquipType.GunOff          => "Aetherotransformer",
-            FullEquipType.Orrery          => "Orrery",
-            FullEquipType.OrreryOff       => "Card Holder",
-            FullEquipType.Katana          => "Katana",
-            FullEquipType.KatanaOff       => "Sheathe",
-            FullEquipType.Rapier          => "Rapier",
-            FullEquipType.RapierOff       => "Focus",
-            FullEquipType.Cane            => "Cane",
-            FullEquipType.Gunblade        => "Gunblade",
-            FullEquipType.Glaives         => "Glaive",
-            FullEquipType.GlaivesOff      => "Glaive (Offhand)",
-            FullEquipType.Scythe          => "Scythe",
-            FullEquipType.Nouliths        => "Nouliths",
-            FullEquipType.Shield          => "Shield",
-            FullEquipType.Saw             => "Saw",
-            FullEquipType.CrossPeinHammer => "Cross Pein Hammer",
-            FullEquipType.RaisingHammer   => "Raising Hammer",
-            FullEquipType.LapidaryHammer  => "Lapidary Hammer",
-            FullEquipType.Knife           => "Round Knife",
-            FullEquipType.Needle          => "Needle",
-            FullEquipType.Alembic         => "Alembic",
-            FullEquipType.Frypan          => "Frypan",
-            FullEquipType.Pickaxe         => "Pickaxe",
-            FullEquipType.Hatchet         => "Hatchet",
-            FullEquipType.FishingRod      => "Fishing Rod",
-            FullEquipType.ClawHammer      => "Clawhammer",
-            FullEquipType.File            => "File",
-            FullEquipType.Pliers          => "Pliers",
-            FullEquipType.GrindingWheel   => "Grinding Wheel",
-            FullEquipType.Awl             => "Awl",
-            FullEquipType.SpinningWheel   => "Spinning Wheel",
-            FullEquipType.Mortar          => "Mortar",
-            FullEquipType.CulinaryKnife   => "Culinary Knife",
-            FullEquipType.Sledgehammer    => "Sledgehammer",
-            FullEquipType.GardenScythe    => "Garden Scythe",
-            FullEquipType.Gig             => "Gig",
-            FullEquipType.Brush           => "Brush",
-            FullEquipType.Palette         => "Palette",
-            FullEquipType.Sabre           => "Sabre",
-            FullEquipType.SabreOff        => "Sabre (Offhand)",
-            FullEquipType.Whip            => "Whip",
-            _                             => "Unknown",
+            FullEquipType.Finger          => "戒指",
+            FullEquipType.Fists           => "格斗武器",
+            FullEquipType.FistsOff        => "格斗武器（副手）",
+            FullEquipType.Sword           => "单手剑",
+            FullEquipType.Axe             => "大斧",
+            FullEquipType.Bow             => "弓",
+            FullEquipType.BowOff          => "箭袋",
+            FullEquipType.Lance           => "长枪",
+            FullEquipType.Staff           => "双手杖",
+            FullEquipType.Wand            => "单手杖",
+            FullEquipType.Book            => "魔导书",
+            FullEquipType.Daggers         => "双剑",
+            FullEquipType.DaggersOff      => "双剑（副手）",
+            FullEquipType.Broadsword      => "双手剑",
+            FullEquipType.Gun             => "火枪",
+            FullEquipType.GunOff          => "以太变换器",
+            FullEquipType.Orrery          => "天球仪",
+            FullEquipType.OrreryOff       => "卡套",
+            FullEquipType.Katana          => "武士刀",
+            FullEquipType.KatanaOff       => "刀鞘",
+            FullEquipType.Rapier          => "刺剑",
+            FullEquipType.RapierOff       => "触媒",
+            FullEquipType.Cane            => "青魔杖",
+            FullEquipType.Gunblade        => "枪刃",
+            FullEquipType.Glaives         => "投掷武器",
+            FullEquipType.GlaivesOff      => "投掷武器（副手）",
+            FullEquipType.Scythe          => "镰刀",
+            FullEquipType.Nouliths        => "贤具",
+            FullEquipType.Shield          => "盾",
+            FullEquipType.Saw             => "刻木工具",
+            FullEquipType.CrossPeinHammer => "锻铁工具",
+            FullEquipType.RaisingHammer   => "铸甲工具",
+            FullEquipType.LapidaryHammer  => "雕金工具",
+            FullEquipType.Knife           => "制革工具",
+            FullEquipType.Needle          => "裁衣工具",
+            FullEquipType.Alembic         => "炼金工具",
+            FullEquipType.Frypan          => "烹调工具",
+            FullEquipType.Pickaxe         => "采矿工具",
+            FullEquipType.Hatchet         => "园艺工具",
+            FullEquipType.FishingRod      => "捕鱼用具",
+            FullEquipType.ClawHammer      => "羊角锤",
+            FullEquipType.File            => "锉刀",
+            FullEquipType.Pliers          => "手钳",
+            FullEquipType.GrindingWheel   => "砂轮机",
+            FullEquipType.Awl             => "平斩",
+            FullEquipType.SpinningWheel   => "纺车",
+            FullEquipType.Mortar          => "研钵",
+            FullEquipType.CulinaryKnife   => "厨刀",
+            FullEquipType.Sledgehammer    => "碎石锤",
+            FullEquipType.GardenScythe    => "园艺镰刀",
+            FullEquipType.Gig             => "渔叉",
+            FullEquipType.Brush           => "笔刷",
+            FullEquipType.Palette         => "调色盘",
+            FullEquipType.Sabre           => "佩刀",
+            FullEquipType.SabreOff        => "佩刀（副手）",
+            FullEquipType.Whip            => "鞭子",
+            _                             => "未知",
         };
 
     /// <summary> Return the actual equipment slot a FullEquipType will be equipped to. </summary>
@@ -484,16 +484,16 @@ public static class FullEquipTypeExtensions
     internal static string OffhandTypeSuffix(this FullEquipType type)
         => type switch
         {
-            FullEquipType.FistsOff   => " (Offhand)",
-            FullEquipType.DaggersOff => " (Offhand)",
-            FullEquipType.GunOff     => " (Aetherotransformer)",
-            FullEquipType.OrreryOff  => " (Card Holder)",
-            FullEquipType.RapierOff  => " (Focus)",
-            FullEquipType.GlaivesOff => " (Offhand)",
-            FullEquipType.BowOff     => " (Quiver)",
-            FullEquipType.KatanaOff  => " (Sheathe)",
-            FullEquipType.SabreOff   => " (Offhand)",
-            FullEquipType.Palette    => " (Palette)",
+            FullEquipType.FistsOff   => "（副手）",
+            FullEquipType.DaggersOff => "（副手）",
+            FullEquipType.GunOff     => "（以太转换器）",
+            FullEquipType.OrreryOff  => "（卡套）",
+            FullEquipType.RapierOff  => "（副手）",
+            FullEquipType.GlaivesOff => "（副手）",
+            FullEquipType.BowOff     => "（箭袋）",
+            FullEquipType.KatanaOff  => "（刀鞘）",
+            FullEquipType.SabreOff   => "（副手）",
+            FullEquipType.Palette    => "（调色盘）",
             _                        => string.Empty,
         };
 
