@@ -1,4 +1,4 @@
-namespace Penumbra.GameData.Enums;
+﻿namespace Penumbra.GameData.Enums;
 
 /// <summary> All available racial scaling parameters. </summary>
 public enum RspAttribute : byte
@@ -47,20 +47,20 @@ public static class RspAttributeExtensions
     public static string ToFullString(this RspAttribute attribute)
         => attribute switch
         {
-            RspAttribute.MaleMinSize   => "Male Minimum Size",
-            RspAttribute.MaleMaxSize   => "Male Maximum Size",
-            RspAttribute.FemaleMinSize => "Female Minimum Size",
-            RspAttribute.FemaleMaxSize => "Female Maximum Size",
-            RspAttribute.BustMinX      => "Bust Minimum X-Axis",
-            RspAttribute.BustMaxX      => "Bust Maximum X-Axis",
-            RspAttribute.BustMinY      => "Bust Minimum Y-Axis",
-            RspAttribute.BustMaxY      => "Bust Maximum Y-Axis",
-            RspAttribute.BustMinZ      => "Bust Minimum Z-Axis",
-            RspAttribute.BustMaxZ      => "Bust Maximum Z-Axis",
-            RspAttribute.MaleMinTail   => "Male Minimum Tail Length",
-            RspAttribute.MaleMaxTail   => "Male Maximum Tail Length",
-            RspAttribute.FemaleMinTail => "Female Minimum Tail Length",
-            RspAttribute.FemaleMaxTail => "Female Maximum Tail Length",
+            RspAttribute.MaleMinSize   => "男性身体最小尺寸",
+            RspAttribute.MaleMaxSize   => "男性身体最大尺寸",
+            RspAttribute.FemaleMinSize => "女性身体最小尺寸",
+            RspAttribute.FemaleMaxSize => "女性身体最大尺寸",
+            RspAttribute.BustMinX      => "胸围最小X轴",
+            RspAttribute.BustMaxX      => "胸围最大X轴",
+            RspAttribute.BustMinY      => "胸围最小Y轴",
+            RspAttribute.BustMaxY      => "胸围最大Y轴",
+            RspAttribute.BustMinZ      => "胸围最小Z轴",
+            RspAttribute.BustMaxZ      => "胸围最大Z轴",
+            RspAttribute.MaleMinTail   => "男性尾巴最小长度",
+            RspAttribute.MaleMaxTail   => "男性尾巴最大长度",
+            RspAttribute.FemaleMinTail => "女性尾巴最小长度",
+            RspAttribute.FemaleMaxTail => "女性尾巴最大长度",
             _                          => throw new InvalidEnumArgumentException(),
         };
 }
