@@ -26,21 +26,21 @@ public class ActorDataDrawer(NameDicts _dicts, DictBNpcNames _bNpcNames) : IGame
     public void Draw()
     {
         DrawBNpcTable();
-        DebugUtility.DrawNameTable("Event NPCs", ref _eNpcFilter,      true, _dicts.ENpcs.Select(kvp => ((ulong)kvp.Key.Id, kvp.Value)));
-        DebugUtility.DrawNameTable("Companions", ref _companionFilter, true, _dicts.Companions.Select(kvp => ((ulong)kvp.Key.Id, kvp.Value)));
-        DebugUtility.DrawNameTable("Mounts",     ref _mountFilter,     true, _dicts.Mounts.Select(kvp => ((ulong)kvp.Key.Id, kvp.Value)));
-        DebugUtility.DrawNameTable("Ornaments",  ref _ornamentFilter,  true, _dicts.Ornaments.Select(kvp => ((ulong)kvp.Key.Id, kvp.Value)));
-        DebugUtility.DrawNameTable("Worlds",     ref _worldFilter,     true, _dicts.Worlds.Select(kvp => ((ulong)kvp.Key.Id, kvp.Value)));
+        DebugUtility.DrawNameTable("事件NPC", ref _eNpcFilter,      true, _dicts.ENpcs.Select(kvp => ((ulong)kvp.Key.Id, kvp.Value)));
+        DebugUtility.DrawNameTable("伙伴", ref _companionFilter, true, _dicts.Companions.Select(kvp => ((ulong)kvp.Key.Id, kvp.Value)));
+        DebugUtility.DrawNameTable("坐骑",     ref _mountFilter,     true, _dicts.Mounts.Select(kvp => ((ulong)kvp.Key.Id, kvp.Value)));
+        DebugUtility.DrawNameTable("装饰",  ref _ornamentFilter,  true, _dicts.Ornaments.Select(kvp => ((ulong)kvp.Key.Id, kvp.Value)));
+        DebugUtility.DrawNameTable("世界",     ref _worldFilter,     true, _dicts.Worlds.Select(kvp => ((ulong)kvp.Key.Id, kvp.Value)));
     }
 
     /// <summary> Draw a list of battle NPCs associated with names and base IDs. </summary>
     private void DrawBNpcTable()
     {
-        using var tree = ImRaii.TreeNode("Battle NPCs");
+        using var tree = ImRaii.TreeNode("战斗NPC");
         if (!tree)
             return;
 
-        var resetScroll = ImGui.InputTextWithHint("##filter", "Filter...", ref _bNpcFilter, 256);
+        var resetScroll = ImGui.InputTextWithHint("##filter", "筛选...", ref _bNpcFilter, 256);
         var height      = ImGui.GetTextLineHeightWithSpacing() + 2 * ImGui.GetStyle().CellPadding.Y;
         using var table = ImRaii.Table("##table", 3, ImGuiTableFlags.RowBg | ImGuiTableFlags.ScrollY | ImGuiTableFlags.BordersOuter,
             new Vector2(-1, 10 * height));
