@@ -1,4 +1,4 @@
-using Dalamud.Interface;
+﻿using Dalamud.Interface;
 using Dalamud.Interface.Utility;
 using Dalamud.Utility;
 using ImGuiNET;
@@ -105,7 +105,7 @@ public class IdentificationDrawer(
         if (!tree)
             return;
 
-        var resetScroll = ImGui.InputTextWithHint("##filter", "Filter...", ref filter, 256);
+        var resetScroll = ImGui.InputTextWithHint("##filter", "筛选...", ref filter, 256);
         ImGui.AlignTextToFramePadding();
         ImGui.TextUnformatted("From");
         ImGui.SameLine(0, ImGui.GetStyle().ItemInnerSpacing.X);
@@ -167,7 +167,7 @@ public class IdentificationDrawer(
         if (!tree)
             return;
 
-        var resetScroll = ImGui.InputTextWithHint("##filter", "Filter...", ref filter, 256);
+        var resetScroll = ImGui.InputTextWithHint("##filter", "筛选...", ref filter, 256);
         var height      = ImGui.GetTextLineHeightWithSpacing() + 2 * ImGui.GetStyle().CellPadding.Y;
         using var table = ImRaii.Table("##table", 2, ImGuiTableFlags.RowBg | ImGuiTableFlags.ScrollY | ImGuiTableFlags.BordersOuter,
             new Vector2(-1, 10 * height));
