@@ -1,3 +1,4 @@
+﻿using System.Collections.Frozen;
 using Penumbra.GameData.Enums;
 
 namespace Penumbra.GameData.Structs;
@@ -213,75 +214,75 @@ public static class Eqp
     public static string ToLocalName(this EqpEntry entry)
         => entry switch
         {
-            EqpEntry.BodyEnabled              => "Enabled",
-            EqpEntry.BodyHideWaist            => "Hide Waist",
-            EqpEntry.BodyHideThighs           => "Hide Thigh Pads",
-            EqpEntry.BodyHideGlovesS          => "Hide Small Gloves",
-            EqpEntry._4                       => "Unknown 4",
-            EqpEntry.BodyHideGlovesM          => "Hide Medium Gloves",
-            EqpEntry.BodyHideGlovesL          => "Hide Large Gloves",
-            EqpEntry.BodyHideGorget           => "Hide Gorget",
-            EqpEntry.BodyShowLeg              => "Show Legs",
-            EqpEntry.BodyShowHand             => "Show Hands",
-            EqpEntry.BodyShowHead             => "Show Head",
-            EqpEntry.BodyShowNecklace         => "Show Necklace",
-            EqpEntry.BodyShowBracelet         => "Show Bracelet",
-            EqpEntry.BodyShowTail             => "Show Tail",
-            EqpEntry.BodyDisableBreastPhysics => "Disable Breast Physics",
+            EqpEntry.BodyEnabled              => "启用",
+            EqpEntry.BodyHideWaist            => "隐藏腰部",
+            EqpEntry.BodyHideThighs           => "隐藏护腿",
+            EqpEntry.BodyHideGlovesS          => "隐藏短手套",
+            EqpEntry._4                       => "未知 Bit4",
+            EqpEntry.BodyHideGlovesM          => "显示中手套",
+            EqpEntry.BodyHideGlovesL          => "显示长手套",
+            EqpEntry.BodyHideGorget           => "隐藏护喉甲",
+            EqpEntry.BodyShowLeg              => "显示腿",
+            EqpEntry.BodyShowHand             => "显示手",
+            EqpEntry.BodyShowHead             => "显示头",
+            EqpEntry.BodyShowNecklace         => "显示项链",
+            EqpEntry.BodyShowBracelet         => "显示手镯",
+            EqpEntry.BodyShowTail             => "显示尾巴",
+            EqpEntry.BodyDisableBreastPhysics => "禁用乳房物理",
             EqpEntry.BodyUsesEvpTable         => "Uses EVP Table",
 
-            EqpEntry.LegsEnabled      => "Enabled",
-            EqpEntry.LegsHideKneePads => "Hide Knee Pads",
-            EqpEntry.LegsHideBootsS   => "Hide Small Boots",
-            EqpEntry.LegsHideBootsM   => "Hide Medium Boots",
-            EqpEntry._20              => "Unknown 20",
-            EqpEntry.LegsShowFoot     => "Show Foot",
-            EqpEntry.LegsShowTail     => "Show Tail",
-            EqpEntry._23              => "Unknown 23",
+            EqpEntry.LegsEnabled      => "启用",
+            EqpEntry.LegsHideKneePads => "隐藏护膝",
+            EqpEntry.LegsHideBootsS   => "隐藏短靴",
+            EqpEntry.LegsHideBootsM   => "隐藏半靴",
+            EqpEntry._20              => "未知 Bit20",
+            EqpEntry.LegsShowFoot     => "显示脚",
+            EqpEntry.LegsShowTail     => "显示尾巴",
+            EqpEntry._23              => "未知 Bit23",
 
-            EqpEntry.HandsEnabled     => "Enabled",
-            EqpEntry.HandsHideElbow   => "Hide Elbow",
-            EqpEntry.HandsHideForearm => "Hide Forearm",
-            EqpEntry._27              => "Unknown 27",
-            EqpEntry.HandShowBracelet => "Show Bracelet",
-            EqpEntry.HandShowRingL    => "Show Left Ring",
-            EqpEntry.HandShowRingR    => "Show Right Ring",
-            EqpEntry._31              => "Unknown 31",
+            EqpEntry.HandsEnabled     => "启用",
+            EqpEntry.HandsHideElbow   => "隐藏肘部",
+            EqpEntry.HandsHideForearm => "隐藏前臂",
+            EqpEntry._27              => "未知 Bit27",
+            EqpEntry.HandShowBracelet => "显示手镯",
+            EqpEntry.HandShowRingL    => "显示左手戒指",
+            EqpEntry.HandShowRingR    => "显示右手戒指",
+            EqpEntry._31              => "未知 Bit31",
 
-            EqpEntry.FeetEnabled   => "Enabled",
-            EqpEntry.FeetHideKnee  => "Hide Knees",
-            EqpEntry.FeetHideCalf  => "Hide Calves",
-            EqpEntry.FeetHideAnkle => "Hide Ankles",
-            EqpEntry._36           => "Unknown 36",
-            EqpEntry._37           => "Unknown 37",
-            EqpEntry._38           => "Unknown 38",
-            EqpEntry._39           => "Unknown 39",
+            EqpEntry.FeetEnabled   => "启用",
+            EqpEntry.FeetHideKnee  => "隐藏膝盖",
+            EqpEntry.FeetHideCalf  => "隐藏小腿",
+            EqpEntry.FeetHideAnkle => "隐藏脚踝",
+            EqpEntry._36           => "未知 Bit36",
+            EqpEntry._37           => "未知 Bit37",
+            EqpEntry._38           => "未知 Bit38",
+            EqpEntry._39           => "未知 Bit39",
 
-            EqpEntry.HeadEnabled           => "Enabled",
-            EqpEntry.HeadHideScalp         => "Hide Scalp",
-            EqpEntry.HeadHideHair          => "Hide Hair",
-            EqpEntry.HeadShowHairOverride  => "Show Hair Override",
-            EqpEntry.HeadHideNeck          => "Hide Neck",
-            EqpEntry.HeadShowNecklace      => "Show Necklace",
-            EqpEntry._46                   => "Unknown 46",
-            EqpEntry.HeadShowEarrings      => "Show Earrings",
-            EqpEntry.HeadShowEarringsHuman => "Show Earrings (Human)",
-            EqpEntry.HeadShowEarringsAura  => "Show Earrings (Au Ra)",
-            EqpEntry.HeadShowEarHuman      => "Show Ears (Human)",
-            EqpEntry.HeadShowEarMiqote     => "Show Ears (Miqo'te)",
-            EqpEntry.HeadShowEarAuRa       => "Show Ears (Au Ra)",
-            EqpEntry.HeadShowEarViera      => "Show Ears (Viera)",
-            EqpEntry._54                   => "Unknown 54",
-            EqpEntry._55                   => "Unknown 55",
-            EqpEntry.HeadShowHrothgarHat   => "Show on Hrothgar",
-            EqpEntry.HeadShowVieraHat      => "Show on Viera",
+            EqpEntry.HeadEnabled           => "启用",
+            EqpEntry.HeadHideScalp         => "隐藏头皮",
+            EqpEntry.HeadHideHair          => "隐藏头发",
+            EqpEntry.HeadShowHairOverride  => "显示头发替代",
+            EqpEntry.HeadHideNeck          => "隐藏颈部",
+            EqpEntry.HeadShowNecklace      => "显示项链",
+            EqpEntry._46                   => "未知 Bit46",
+            EqpEntry.HeadShowEarrings      => "显示耳环",
+            EqpEntry.HeadShowEarringsHuman => "显示耳环(人族)",
+            EqpEntry.HeadShowEarringsAura  => "显示耳环(敖龙族)",
+            EqpEntry.HeadShowEarHuman      => "显示耳朵(人族)",
+            EqpEntry.HeadShowEarMiqote     => "显示耳朵(猫魅族)",
+            EqpEntry.HeadShowEarAuRa       => "显示耳朵(敖龙族)",
+            EqpEntry.HeadShowEarViera      => "显示耳朵(维埃拉族)",
+            EqpEntry._54                   => "头部未知头盔1 Bit54",
+            EqpEntry._55                   => "头部未知头盔2 Bit55",
+            EqpEntry.HeadShowHrothgarHat   => "显示硌狮族帽子",
+            EqpEntry.HeadShowVieraHat      => "显示维埃拉族帽子",
             EqpEntry.HeadUsesEvpTable      => "Uses EVP Table",
 
-            EqpEntry._59 => "Unknown 59",
-            EqpEntry._60 => "Unknown 60",
-            EqpEntry._61 => "Unknown 61",
-            EqpEntry._62 => "Unknown 62",
-            EqpEntry._63 => "Unknown 63",
+            EqpEntry._59 => "未知 Bit59",
+            EqpEntry._60 => "未知 Bit60",
+            EqpEntry._61 => "未知 Bit61",
+            EqpEntry._62 => "未知 Bit62",
+            EqpEntry._63 => "未知 Bit63",
 
             _ => throw new InvalidEnumArgumentException(),
         };
@@ -295,28 +296,28 @@ public static class Eqp
     }
 
     /// <summary> A list of flags affecting the body. </summary>
-    public static readonly IReadOnlyList<EqpEntry> EqpAttributesBody  = GetEntriesForSlot(EquipSlot.Body);
+    public static readonly IReadOnlyList<EqpEntry> EqpAttributesBody = GetEntriesForSlot(EquipSlot.Body);
 
     /// <summary> A list of flags affecting the legs. </summary>
-    public static readonly IReadOnlyList<EqpEntry> EqpAttributesLegs  = GetEntriesForSlot(EquipSlot.Legs);
+    public static readonly IReadOnlyList<EqpEntry> EqpAttributesLegs = GetEntriesForSlot(EquipSlot.Legs);
 
     /// <summary> A list of flags affecting the hands. </summary>
     public static readonly IReadOnlyList<EqpEntry> EqpAttributesHands = GetEntriesForSlot(EquipSlot.Hands);
 
     /// <summary> A list of flags affecting the feet. </summary>
-    public static readonly IReadOnlyList<EqpEntry> EqpAttributesFeet  = GetEntriesForSlot(EquipSlot.Feet);
+    public static readonly IReadOnlyList<EqpEntry> EqpAttributesFeet = GetEntriesForSlot(EquipSlot.Feet);
 
     /// <summary> A list of flags affecting the head. </summary>
-    public static readonly IReadOnlyList<EqpEntry> EqpAttributesHead  = GetEntriesForSlot(EquipSlot.Head);
+    public static readonly IReadOnlyList<EqpEntry> EqpAttributesHead = GetEntriesForSlot(EquipSlot.Head);
 
-    // TODO: FrozenDictionary
     /// <summary> A dictionary mapping slots to a list of their affected flags. </summary>
-    public static readonly IReadOnlyDictionary<EquipSlot, IReadOnlyList<EqpEntry>> EqpAttributes = new Dictionary<EquipSlot, IReadOnlyList<EqpEntry>>
-    {
-        [EquipSlot.Body]  = EqpAttributesBody,
-        [EquipSlot.Legs]  = EqpAttributesLegs,
-        [EquipSlot.Hands] = EqpAttributesHands,
-        [EquipSlot.Feet]  = EqpAttributesFeet,
-        [EquipSlot.Head]  = EqpAttributesHead,
-    };
+    public static readonly IReadOnlyDictionary<EquipSlot, IReadOnlyList<EqpEntry>> EqpAttributes =
+        FrozenDictionary.ToFrozenDictionary<EquipSlot, IReadOnlyList<EqpEntry>>(
+        [
+            new KeyValuePair<EquipSlot, IReadOnlyList<EqpEntry>>(EquipSlot.Body,  EqpAttributesBody),
+            new KeyValuePair<EquipSlot, IReadOnlyList<EqpEntry>>(EquipSlot.Legs,  EqpAttributesLegs),
+            new KeyValuePair<EquipSlot, IReadOnlyList<EqpEntry>>(EquipSlot.Hands, EqpAttributesHands),
+            new KeyValuePair<EquipSlot, IReadOnlyList<EqpEntry>>(EquipSlot.Feet,  EqpAttributesFeet),
+            new KeyValuePair<EquipSlot, IReadOnlyList<EqpEntry>>(EquipSlot.Head,  EqpAttributesHead),
+        ]);
 }
